refactor: move code logic to abstraction layer to avoid duplicate code
